move some code to reduce large file size
